feat : roomcheck page (roomopen/:id)의 userInfo GET API연결 및 구현 완료
