+++ pages/oauth/kakao.tsx
@@ -22,6 +22,7 @@ function kakao() {
 			}
 			sendCode(code);
 		}
+		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [])
 }
